Moved file location and cleaned up a few spots
